Introduce inheritance to Graphene python worker class
- Create a base python worker
- Make existing worker to derive from base and rename it as SingletonPythonWorker
- Create a new class WPEPythonWorker deriving from base worker
- Update Makefile to handle both types of workers
